+++ src/SpidersMessageHeader.jl
@@ -6,7 +6,7 @@ struct SpidersMessageHeader{T<:AbstractArray{UInt8}}
     buffer::T
     offset::Int64
     acting_version::UInt16
-    function SpidersMessageHeader(buffer::T, offset::Int64=0, acting_version::Integer=0) where {T}
+    function SpidersMessageHeader(buffer::T, offset::Integer=0, acting_version::Integer=0) where {T}
         new{T}(buffer, offset, acting_version)
     end
 end
